rename sub_commands files to subcommands
